fix: convert files to ts and tsx
