+++ package.json
@@ -44,6 +44,7 @@
   },
   "devDependencies": {
     "@types/node": "18.11.12",
+    "@types/react-gtm-module": "^2.0.1",
     "cypress": "^12.2.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-import": "^2.27.5",
