+++ pages/_app.tsx
@@ -1,26 +1,31 @@
 import '../styles/global.css';
 
-import { CacheProvider } from '@emotion/react';
+import { CacheProvider, EmotionCache } from '@emotion/react';
 import { CssBaseline, ThemeProvider } from '@mui/material';
 import Head from 'next/head';
 import { useRouter } from 'next/router';
-import PropTypes from 'prop-types';
 import React, { useEffect } from 'react';
 import TagManager from 'react-gtm-module';
 
 import Footer from '../components/layout/Footer.tsx';
 import Header from '../components/layout/Header.tsx';
 import theme from '../styles/themes/theme.tsx';
-import createEmotionCache from '../utils/createEmotionCache';
+import createEmotionCache from '../utils/createEmotionCache.ts';
 
 const clientSideEmotionCache = createEmotionCache();
-const gtmId = process.env.NEXT_PUBLIC_GTM_ID;
+const gtmId = String(process.env.NEXT_PUBLIC_GTM_ID);
+
+interface AppProps {
+  Component: React.ComponentType;
+  emotionCache?: EmotionCache;
+  pageProps: any;
+}
 
 function MyApp({
   Component,
   pageProps,
   emotionCache = clientSideEmotionCache,
-}) {
+}: AppProps) {
   const router = useRouter();
 
   useEffect(() => {
@@ -53,13 +58,5 @@ function MyApp({
     </CacheProvider>
   );
 }
-
-MyApp.propTypes = {
-  Component: PropTypes.elementType.isRequired,
-  // eslint-disable-next-line react/forbid-prop-types
-  emotionCache: PropTypes.any,
-  // eslint-disable-next-line react/forbid-prop-types
-  pageProps: PropTypes.any,
-};
 
 export default MyApp;
